requirements.txt
playwright>=1.44.0
pyyaml>=6.0
python-dateutil>=2.9
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
pillow>=10.0
